+++ hfie/endpoint.py
@@ -1,10 +1,8 @@
-import json
-
 import requests
 import typer
 
 from hfie.settings import Settings
-from hfie.utils import format_table
+from hfie.utils import format_table, load_json
 
 settings = Settings()
 
@@ -14,6 +12,7 @@ headers = {
     "Authorization": f"Bearer {settings.token}",
     "Content-Type": "application/json",
 }
+API_ERROR_MESSAGE = "An error occured while making the API call"
 
 
 @app.command()
@@ -21,13 +20,13 @@ def list(json: bool = typer.Option(False, help="Prints the full output in JSON."
     """
     List all the deployed endpoints
     """
-    r = requests.get(f"{settings.endpoint_url}", headers=headers)
+    response = requests.get(settings.endpoint_url, headers=headers)
 
     if json:
-        return typer.echo(r.json())
+        return typer.echo(response.json())
 
     else:
-        data = r.json()
+        data = response.json()
 
         if data.get("items"):
 
@@ -64,11 +63,29 @@ def create(
 ):
     """
     Create an endpoint
+
+    Args:
+        data (str): Path to JSON data to create the endpoint
     """
-    with open(data) as f:
-        data = json.load(f)
-    r = requests.post(f"{settings.endpoint_url}", headers=headers, json=data)
-    typer.echo(r.json())
+
+    data = load_json(data)
+    try:
+        response = requests.post(settings.endpoint_url, headers=headers, json=data)
+        response.raise_for_status()
+    except requests.exceptions.HTTPError as e:
+        if response.json().get("error"):
+            typer.secho(
+                f"Error creating endpoint: {response.json()['error']}",
+                fg=typer.colors.RED,
+            )
+        else:
+            typer.secho("Error creating endpoint", fg=typer.colors.RED)
+        raise SystemExit(e)
+    except requests.exceptions.RequestException as e:
+        typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+        raise SystemExit(e)
+    typer.secho("Endpoint created successfully", fg=typer.colors.GREEN)
+    typer.echo(response.json())
 
 
 @app.command()
@@ -79,10 +96,27 @@ def update(
     """
     Update an endpoint
     """
-    with open(data) as f:
-        data = json.load(f)
-    r = requests.put(f"{settings.endpoint_url}/{name}", headers=headers, json=data)
-    typer.echo(r.json())
+    data = load_json(data)
+
+    try:
+        response = requests.put(
+            f"{settings.endpoint_url}/{name}", headers=headers, json=data
+        )
+    except requests.exceptions.HTTPError as e:
+        if response.json().get("error"):
+            typer.secho(
+                f"Error updating endpoint: {response.json()['error']}",
+                fg=typer.colors.RED,
+            )
+        else:
+            typer.secho("Error updating endpoint", fg=typer.colors.RED)
+        raise SystemExit(e)
+    except requests.exceptions.RequestException as e:
+        typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+        raise SystemExit(e)
+
+    typer.secho("Endpoint updated successfully", fg=typer.colors.GREEN)
+    typer.echo(response.json())
 
 
 @app.command()
@@ -95,24 +129,27 @@ def delete(
     """
     Delete an endpoint
     """
+
     if not force:
         delete_endpoint = typer.confirm(
-            f"Are you sure you want to delete endpoint. Use --force to override"
+            "Are you sure you want to delete endpoint. Use --force to override"
         )
+
         if not delete_endpoint:
             typer.echo("Not deleting endpoint")
             raise typer.Abort()
 
     if force or delete_endpoint:
-        r = requests.delete(f"{settings.endpoint_url}/{name}", headers=headers)
-        typer.echo(r.json())
+        try:
+            response = requests.delete(
+                f"{settings.endpoint_url}/{name}", headers=headers, json={}
+            )
+            response.raise_for_status()
+        except requests.exceptions.RequestException as e:
+            typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+            raise SystemExit(e)
 
-
-def get_info(name: str):
-
-    r = requests.get(f"{settings.endpoint_url}/{name}", headers=headers)
-
-    return r.json()
+        typer.secho("Endpoint deleted successfully", fg=typer.colors.GREEN)
 
 
 @app.command()
@@ -123,7 +160,17 @@ def info(
     """
     Get info about an endpoint
     """
-    info = get_info(name)
+
+    try:
+        response = requests.get(
+            f"{settings.endpoint_url}/{name}", headers=headers, json={}
+        )
+        response.raise_for_status()
+    except requests.exceptions.RequestException as e:
+        typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+        raise SystemExit(e)
+
+    info = response.json()
 
     if info.get("name"):
 
@@ -162,21 +209,61 @@ def logs(name: str = typer.Argument(..., help="Endpoint name")):
     """
     Get logs about an endpoint
     """
-    r = requests.get(f"{settings.endpoint_url}/{name}/logs", headers=headers)
-    typer.echo(r.content)
+    try:
+        response = requests.get(
+            f"{settings.endpoint_url}/{name}/logs", headers=headers, json={}
+        )
+        response.raise_for_status()
+    except requests.exceptions.RequestException as e:
+        typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+        raise SystemExit(e)
+
+    typer.echo(response.content)
 
 
 @app.command()
 def test(
     name: str = typer.Argument(..., help="Endpoint name"),
-    inputs: str = typer.Argument(..., help="Input to send the model."),
+    inputs: str = typer.Argument(None, help="Input to send the model."),
+    input_file: str = typer.Option(
+        None, help="Path to JSON file containing queries to send to the model."
+    ),
 ):
     """
     Test an endpoint
     """
 
-    info = get_info(name)
+    if not inputs and not input_file:
+        typer.secho(
+            "You must provide either an input string or an input JSON containing your queries",
+            fg=typer.colors.RED,
+        )
+        raise typer.Abort()
+
+    # Get the endpoint url from endpoint info
+    try:
+        response = requests.get(
+            f"{settings.endpoint_url}/{name}", headers=headers, json={}
+        )
+        response.raise_for_status()
+    except requests.exceptions.RequestException as e:
+        typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+        raise SystemExit(e)
+
+    info = response.json()
     url = info["status"]["url"]
-    data = {"inputs": inputs, "parameters": {"top_k": 10}}
-    r = requests.post(url, headers=headers, json=data)
-    typer.echo(r.json())
+
+    if input_file:
+        data = load_json(input_file)
+    else:
+        data = {"inputs": inputs, "parameters": {"top_k": 10}}
+
+    # Send a call to the endpoint
+    try:
+        response = requests.post(url, headers=headers, json=data)
+        response.raise_for_status()
+    except requests.exceptions.RequestException as e:
+        typer.secho(API_ERROR_MESSAGE, fg=typer.colors.RED)
+        raise SystemExit(e)
+
+    typer.echo(response.json())